refactor(app-deployment) :: AppRevisionResponseDto max_memory 타입을 Double로 변경

+++ src/main/java/madp/appdeployment/domain/infrastructure/client/response/AppRevisionResponseDto.java
@@ -10,7 +10,7 @@ public record AppRevisionResponseDto(
         Integer maxCpu,
 
         @JsonProperty("max_memory")
-        Integer maxMemory,
+        Double maxMemory,
 
         @JsonProperty("max_disk")
         Integer maxDisk,
